+++ packages/dht/test/unit/RecursiveOperationManager.test.ts
@@ -13,8 +13,8 @@ import {
 import { RecursiveOperationManager } from '../../src/dht/recursive-operation/RecursiveOperationManager'
 import { LocalDataStore } from '../../src/dht/store/LocalDataStore'
 import { v4 } from 'uuid'
-import { MockRouter } from '../utils/mock/Router'
-import { MockTransport } from '../utils/mock/Transport'
+import { MockRouter } from '../utils/mock/MockRouter'
+import { MockTransport } from '../utils/mock/MockTransport'
 import { FakeRpcCommunicator } from '../utils/FakeRpcCommunicator'
 import { Router } from '../../src/dht/routing/Router'
 import { ITransport } from '../../src/transport/ITransport'
