+++ packages/trackerless-network/test/unit/TemporaryConnectionRpcLocal.test.ts
@@ -1,6 +1,6 @@
 import { StreamPartIDUtils } from '@streamr/protocol'
 import { TemporaryConnectionRpcLocal } from '../../src/logic/temporary-connection/TemporaryConnectionRpcLocal'
-import { MockTransport } from '../utils/mock/Transport'
+import { MockTransport } from '../utils/mock/MockTransport'
 import { createMockPeerDescriptor } from '../utils/utils'
 import { ListeningRpcCommunicator, getDhtAddressFromRaw } from '@streamr/dht'
